use location child component, lazy-load store - WIP
