+++ app/frontend/src/components/admin/inspection/LocationEdit.vue
@@ -14,173 +14,10 @@
         </v-card-title>
         <v-card-text>
           <v-container class="pa-0">
-            <v-row>
-              <v-col>
-                <h4>Accomodation details</h4>
-              </v-col>
-            </v-row>
 
-            <v-row>
-              <v-col cols="12" sm="6" lg="5">
-                <v-menu
-                  v-model="startDateMenu"
-                  :close-on-content-click="true"
-                  :nudge-right="40"
-                  transition="scale-transition"
-                  offset-y
-                  min-width="290px"
-                >
-                  <template v-slot:activator="{ on }">
-                    <label>Operation Start Date</label>
-                    <v-text-field
-                      v-model="startDate"
-                      :rules="startDateRules"
-                      placeholder="yyyy-mm-dd"
-                      append-icon="event"
-                      v-on:click:append="startDateMenu=true"
-                      readonly
-                      v-on="on"
-                      dense
-                      flat
-                      outlined
-                      solo
-                    ></v-text-field>
-                  </template>
-                  <v-date-picker v-model="startDate" @input="startDateMenu = false"></v-date-picker>
-                </v-menu>
-              </v-col>
+            <!-- child Location form component -->
+            <Location :reviewMode="reviewMode"/>
 
-              <v-col cols="12" sm="6" lg="5">
-                <v-menu
-                  v-model="endDateMenu"
-                  :close-on-content-click="true"
-                  :nudge-right="40"
-                  transition="scale-transition"
-                  offset-y
-                  min-width="290px"
-                >
-                  <template v-slot:activator="{ on }">
-                    <label>Operation End Date</label>
-                    <v-text-field
-                      v-model="endDate"
-                      :rules="endDateRules"
-                      placeholder="yyyy-mm-dd"
-                      append-icon="event"
-                      v-on:click:append="endDateMenu=true"
-                      readonly
-                      v-on="on"
-                      dense
-                      flat
-                      outlined
-                      solo
-                    ></v-text-field>
-                  </template>
-                  <v-date-picker v-model="endDate" @input="endDateMenu = false"></v-date-picker>
-                </v-menu>
-              </v-col>
-            </v-row>
-
-            <v-row>
-              <v-col cols="12" sm="6" lg="5">
-                <label>Closest Community / Town / City</label>
-                <CityLookup
-                  :city-field-model.sync="locationCity"
-                  :city-latitude-field-model.sync="cityLatitude"
-                  :city-longitude-field-model.sync="cityLongitude"
-                  :field-rules="locationCityRules"
-                  :fieldModel="locationCity"
-                />
-                <v-text-field v-model="cityLatitude" class="d-none" />
-                <v-text-field v-model="cityLongitude" class="d-none" />
-              </v-col>
-            </v-row>
-
-            <v-row>
-              <v-col cols="12" sm="4" lg="3">
-                <label>Number of workers at this location</label>
-                <v-text-field
-                  v-model="numberOfWorkers"
-                  :rules="numberOfWorkersRules"
-                  type="number"
-                  min="1"
-                  dense
-                  flat
-                  outlined
-                  solo
-                />
-              </v-col>
-            </v-row>
-
-            <h4>Type of accommodation for workers at this location (check all that apply)</h4>
-
-            <v-checkbox v-model="accTents" label="Tents near worksite"></v-checkbox>
-
-            <div v-if="accTents">
-              <v-row>
-                <v-col cols="12" lg="10">
-                  <label>
-                    Details (eg:
-                    <em>"1km from HWY 1 at 100 mile house north on Logging Road"</em>)
-                  </label>
-                  <v-text-field v-model="tentDetails" dense flat outlined solo />
-                </v-col>
-              </v-row>
-            </div>
-
-            <v-checkbox v-model="accMotel" label="Motel / Hotel in town"></v-checkbox>
-            <div v-if="accMotel">
-              <v-row>
-                <v-col cols="12" sm="6" lg="5">
-                  <label>Name</label>
-                  <v-text-field v-model="motelName" dense flat outlined solo />
-                </v-col>
-              </v-row>
-
-              <v-row>
-                <v-col cols="12" sm="6" lg="5">
-                  <label>Address line 1</label>
-                  <v-text-field v-model="motelAddressLine1" dense flat outlined solo />
-                </v-col>
-
-                <v-col cols="12" sm="6" lg="5">
-                  <label>Address line 2 (Optional)</label>
-                  <v-text-field v-model="motelAddressLine2" dense flat outlined solo />
-                </v-col>
-              </v-row>
-
-              <v-row>
-                <v-col cols="12" sm="6" lg="5">
-                  <label>City</label>
-                  <v-text-field v-model="motelCity" dense flat outlined solo />
-                </v-col>
-                <v-col cols="12" sm="3" lg="2">
-                  <label>Province</label>
-                  <v-select
-                    dense
-                    flat
-                    outlined
-                    solo
-                    single-line
-                    label="select"
-                    v-model="motelProvince"
-                    :items="provinces"
-                  />
-                </v-col>
-              </v-row>
-
-              <v-row>
-                <v-col cols="12" sm="3" lg="2">
-                  <label>Postal Code</label>
-                  <v-text-field v-model="motelPostalCode" dense flat outlined solo />
-                </v-col>
-              </v-row>
-            </div>
-
-            <v-row>
-              <v-col>
-                <v-checkbox v-model="accWorkersHome" label="Worker's home in community"></v-checkbox>
-              </v-col>
-            </v-row>
           </v-container>
         </v-card-text>
         <v-card-actions class="pa-3 pb-10">
@@ -193,16 +30,46 @@
 </template>
 
 <script>
-import { mapActions, mapGetters, mapMutations } from 'vuex';
+import Vuex from 'vuex';
+import { mapGetters } from 'vuex';
 
-import CityLookup from '@/components/form/CityLookup.vue';
-
+import Location from '@/components/form/Location.vue';
 import { SilvipcRoles } from '@/utils/constants';
+
+/* eslint-disable */
+import ipcService from '../../../services/ipcService';
+
+const localStore = new Vuex.Store({
+  state: {
+    // copy location state from global store
+    location: this.location
+  },
+  actions: {
+    async updateForm({ commit, state }) {
+
+      console.log(localStore.location);
+
+      try {
+        const body = {
+          location: localStore.location
+        };
+        const response = await ipcService.updateIPCContent(body);
+
+        if (!response.data) {
+          throw new Error('No response data from API while submitting update');
+        }
+      } catch (error) {
+        console.error(`Error submitting update: ${error}`); // eslint-disable-line no-console
+      }
+    }
+
+  }
+})
 
 export default {
   name: 'LocationEdit',
   components: {
-    CityLookup,
+    Location,
   },
   props: {
     ipcPlanId: {
@@ -215,26 +82,9 @@ export default {
       showEditBtn: false,
       dialog: false,
       updateValid: false,
+      reviewMode: false,
       startDateMenu: false,
       endDateMenu: false,
-
-      // Todo: constants file
-      provinces: ['AB','BC','MB','NB','NL','NS','NT','NU','ON','PE','QC','SK','YT'],
-      startDateRules: [
-        v => !!v || 'Start date is required'
-      ],
-      endDateRules: [
-        v => !!v || 'End date is required'
-      ],
-      locationCityRules: [
-        v => !!v || 'Closest Community / Town / City is required'
-      ],
-      // Todo, put in some utility fxn somewhere if needed again
-      numberOfWorkersRules: [
-        v => (new RegExp('^[-+]?\\d+$')).test(v) || 'invalid # of workers',
-        v => v > 0 || '# of workers must be greater than 0',
-        v => v < 9999 || '# of workers must 9999 or less'
-      ],
     };
   },
   computed: {
@@ -249,74 +99,54 @@ export default {
     // location data
     startDate: {
       get() { return this.location.startDate; },
-      set(value) { this.updateLocation({['startDate']: value}); }
     },
     endDate: {
       get() { return this.location.endDate; },
-      set(value) { this.updateLocation({['endDate']: value}); }
     },
     locationCity: {
       get() { return this.location.city; },
-      set(value) { this.updateLocation({['city']: value}); }
     },
     cityLatitude: {
       get() { return this.location.cityLatitude; },
-      set(value) { this.updateLocation({['cityLatitude']: value}); }
     },
     cityLongitude: {
       get() { return this.location.cityLongitude; },
-      set(value) { this.updateLocation({['cityLongitude']: value}); }
     },
     numberOfWorkers: {
       get() { return this.location.numberOfWorkers ? this.location.numberOfWorkers.toString() : ''; },
-      set(value) { this.updateLocation({['numberOfWorkers']:
-        Number.isNaN(value) ? 0 : Number.parseInt(value)});
-      }
     },
     accTents: {
       get() { return this.location.accTents; },
-      set(value) { this.updateLocation({['accTents']: value}); }
     },
     tentDetails: {
       get() { return this.location.tentDetails; },
-      set(value) { this.updateLocation({['tentDetails']: value}); }
     },
     accMotel: {
       get() { return this.location.accMotel; },
-      set(value) { this.updateLocation({['accMotel']: value}); }
     },
     motelName: {
       get() { return this.location.motelName; },
-      set(value) { this.updateLocation({['motelName']: value}); }
     },
     motelAddressLine1: {
       get() { return this.location.motelAddressLine1; },
-      set(value) { this.updateLocation({['motelAddressLine1']: value}); }
     },
     motelAddressLine2: {
       get() { return this.location.motelAddressLine2; },
-      set(value) { this.updateLocation({['motelAddressLine2']: value}); }
     },
     motelCity: {
       get() { return this.location.motelCity; },
-      set(value) { this.updateLocation({['motelCity']: value}); }
     },
     motelProvince: {
       get() { return this.location.motelProvince; },
-      set(value) { this.updateLocation({['motelProvince']: value}); }
     },
     motelPostalCode: {
       get() { return this.location.motelPostalCode; },
-      set(value) { this.updateLocation({['motelPostalCode']: value}); }
     },
     accWorkersHome: {
       get() { return this.location.accWorkersHome; },
-      set(value) { this.updateLocation({['accWorkersHome']: value}); }
     },
   },
   methods: {
-    ...mapMutations('form', ['updateLocation']),
-    ...mapActions('form', ['updateForm']),
 
     async update() {
       if(this.$refs.form.validate()) {
@@ -339,7 +169,20 @@ export default {
   mounted() {
     // show edit button for inspectors on submission review page
     this.showEditBtn = (this.$route.name == 'Submission' && this.hasSilvipcRoles([SilvipcRoles.INSPECTOR])) ? true : false;
-  }
+  },
+  created() {
+    // lazy-load global store
+    if(this.$store.hasModule('localStore')) {
+      this.$store.unregisterModule('localStore');
+    }
+    this.$store.registerModule('localStore', localStore);
+  },
+  beforeDestroy() {
+    // unload store
+    this.$store.unregisterModule('localStore');
+  },
+
+
 };
 </script>
